DataSet - Extract class responsible for file input
